Added api bridge and code refactoring
Added api bridge and performed some code refactoring.
