+++ create.html
@@ -2,7 +2,7 @@
     <head>
         <script type="module">
             import {Vinyl} from "./js/vinyl.js";
-            import {ERROR} from "./js/vinyl-frontend.js";
+            import {ERROR} from "./js/codes.js";
 
             let vinyl = new Vinyl();
             vinyl.renderForm("new_vinyl");
