+++ edit.html
@@ -2,7 +2,8 @@
     <head>
         <script type="module">
             import {Vinyl} from "./js/vinyl.js";
-            import {fetchVinylById, getVinlyId, ERROR} from "./js/vinyl-frontend.js";
+            import {fetchVinylById, getVinlyId} from "./js/vinyl-frontend.js";
+            import {ERROR} from "./js/codes.js";
 
             try{
                 const vinylId = getVinlyId();
