+++ index.html
@@ -2,7 +2,8 @@
     <head>
         <script type="module">
         import {VinylLineItem} from "./js/vinyl.js";
-        import {fetchVinylsByFilter, ERROR} from "./js/vinyl-frontend.js";
+        import {fetchVinylsByFilter} from "./js/vinyl-frontend.js";
+        import {ERROR} from "./js/codes.js";
 
         let filter = "";
