+++ js/codes.js
@@ -1,0 +1,5 @@
+export const ERROR = {
+    no_id_specified: "_no_id_specified",
+    no_id_matched: "_no_match_found_for_id",
+    request_failed: "_request_failed",
+};
